add forestjs module into browser
