+++ forest/templates/index.html
@@ -14,6 +14,7 @@
 {% block postamble %}
 <link rel="stylesheet" href="forest/static/style.css" type="text/css" media="all">
 <script src="forest/static/script.js" charset="utf-8"></script>
+<script src="forest/static/forest-min.js" charset="utf-8"></script>
 {% endblock %}
 
 {% block contents %}
